% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rgeyer-package}
\alias{rgeyer}
\alias{rgeyer-package}
\title{Simulate Geyer Saturation Model}
\description{
Simulate Geyer Saturation Model
}
\details{
Some functions for dealing with Geyer's saturation model. Can do multi-type with type-to-type pontentials,
and multi-scale in step-function fashion.

Simulate univariate: \code{\link{rstepper}}

Simulate multivariate: \code{\link{rstepper_multi}}
}
