% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper-residuals.R
\name{rgeyer_resid}
\alias{rgeyer_resid}
\title{Residuals for the multivariate multirange saturation process}
\usage{
rgeyer_resid(x, d)
}
\description{
Residuals for the multivariate multirange saturation process
}

