% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrelative-stepper.R
\name{rrelativestepper}
\alias{rrelativestepper}
\title{Simulate the univariate relative saturation model}
\usage{
rrelativestepper(theta, R, K, r, n = NULL, bbox = cbind(0:1, 0:1),
  iter = 10, x0 = NULL, toroidal = 0, dbg = 0, trend = NULL)
}
\arguments{
\item{theta}{weights, length K+1, first element controls intensity}

\item{R}{maximum intereaction}

\item{K}{number of steps}

\item{r}{alternative to R,K. Should be increasing, not containing 0. Length will be taken as K.}

\item{n}{alternative to theta0: fixed point counts. Then theta should be K-length. see details.}

\item{bbox}{bounding box}

\item{iter}{iteration of bd}

\item{x0}{the starting pattern location matrix}

\item{toroidal}{toroidal distances?}

\item{dbg}{dbg}

\item{trend}{Optional im-object defining a trend.}
}
\description{
Simulate a saturation with K different annulus and saturation relative to point count.
}
\details{
If r not given, the steps will be (1:K)*R/K. 0 will be the beginning of the first interval,
so theta[2] is mapped to range (0,r[1]).
If trend is given, it should be a 'spatstat' im-object covering the bbox. Will not be checked (at the moment).

The saturation of the original Geyer's model is a fixed constant c>0, so that each point contributes
for each annuli k

\deqn{t_k(u,x) = min[c, n_k(u,x)]}

where \eqn{n_k(u,x)} is the neighbour count in the annuli

\deqn{A_k=b(u,{r_k})-b(u,r_{k-1}) .}

The model here uses

\deqn{t_k(u,x) = min[c_k(x), n_k(u,x)] / c_k(x)}

where we use the expected count under Poisson model to adjust saturation, i.e.

\deqn{c_k(x) = ceiling( E_0 n(u,x) ) = ceiling[|x|/|W| |A_k|] = ceiling[\lambda|A_k|]}

Here \eqn{\lambda} is the intensity observed of the process. This way the process stays stable i.e.

\deqn{t_k(u,x) = 0,1/c_k(x)..., 1 < \infty}
}

