% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstepper-univ.R
\name{rstepper}
\alias{rstepper}
\title{Simulate the univariate Step-Geyer model}
\usage{
rstepper(theta, R, K, r, n = NULL, bbox = cbind(0:1, 0:1), iter = 10,
  x0 = NULL, toroidal = 0, dbg = 0, trend = NULL)
}
\arguments{
\item{theta}{weights, length K+1, first element controls intensity}

\item{R}{maximum intereaction}

\item{K}{number of steps}

\item{r}{alternative to R,K. Should be increasing, not containing 0. Length will be taken as K.}

\item{n}{alternative to theta0: fixed point counts. Then theta should be K-length. see details.}

\item{bbox}{bounding box}

\item{iter}{iteration of bd}

\item{x0}{the starting pattern location matrix}

\item{toroidal}{toroidal distances?}

\item{dbg}{dbg}

\item{trend}{Optional im-object defining a trend.}
}
\description{
Simulate Geyer's saturation with K different annulus and saturation fixed to 1.
}
\details{
If r not given, the steps will be (1:K)*R/K. 0 will be the beginning of the first interval, so theta[2] is mapped to range (0,r[1]).
If trend is given, it should be a 'spatstat' im-object covering the bbox. Will not be checked (at the moment).
}

