% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstepper-papangelou-evaluate.R
\name{rstepper_papangelou}
\alias{rstepper_papangelou}
\title{Compute the papangelou conditional intensity for multi-type multi-range saturation model}
\usage{
rstepper_papangelou(x, m, at = "data", grid_dimyx, bbox, theta0, theta1,
  theta2, covariates)
}
\arguments{
\item{x}{the point pattern data location matrix}

\item{m}{mark vector, converted to 0-based integers by as.integer(factor(m))-1}

\item{at}{which ones to compute: "data": at data points, "grid": at grid locations}

\item{grid_dimyx}{if at="grid", the (nrow, ncol) dimensions of the grid}

\item{bbox}{if at="grid", the bounding box for the grid}

\item{theta0}{intensity parameters, see Details.}

\item{theta1}{intra-type parameters}

\item{theta2}{inter-type parameters}

\item{covariates}{an optional list of im-objects}
}
\description{
Compute the papangelou conditional intensity for multi-type multi-range saturation model
}
\details{
The parametrisation is much like in \link{rstepper-multi}. Let's say nlevels(factor(m)) = p, and length(covariates)=k:

theta0: list(c(intercept_1, cov_11, cov_12, ..., cov_1k), ..., c(intercept_p, ..., cov_pk)) with individual coefficients being real numbers.

theta1: list(list(theta1_1, r1_1, c1_1), ..., c(theta1_p, r1_p, c1_p)) where theta1_i are the coefficients for the step functions defined by ranges r1_1, saturation levels per step in c1_i. So length(theta1) = p.

theta2: like theta1 for each pair i=(l,k), l < k in 1,...,p. Length(theta2) = p*(p-1)/2.

covariates: list of im-objects, length(covariates) = k. If missing, the cov_il coefficients in theta0 should be omitted.

If covariates are not given, theta0 can be a simple vector of length p.
}
