% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper_biv-papangelou-components.R
\name{stepper_biv_components}
\alias{stepper_biv_components}
\title{Compute the components of papangelou for the bivariate stepper model}
\usage{
stepper_biv_components(from, to, r, bbox, dbg = 0, toroidal = 0)
}
\arguments{
\item{from}{coordinates of the data, third column 0/1 marks}

\item{to}{coordinate matrix for locations for evaluation of papangelou. Assumed to be different from 'from'}

\item{r}{alternative to K and R, the r vector itself. Increasing, 0 shouldn't be included.}

\item{bbox}{bounding box}

\item{dbg}{dbg level}

\item{toroidal}{>0 use toroidal distances}
}
\value{
Matrix with dimension (nrow(to), K) with
d_k(to_i|from) = 1( ne_k(to_i, from[!mark_i])>0) with ne_k(to_i, from[!mark_i]>0) = sum_j 1(mark_j!=mark_i)(r[k-1] < ||to_i - from_j|| < r[k])

If 'to' is NULL, computes the components for 'from' locations, i.e. the papangelou for each data point.
}
\description{
Compute the components of papangelou for the bivariate stepper model
}

