% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper-papangelou-components.R
\name{stepper_components}
\alias{stepper_components}
\title{Compute the components of papangelou for the stepper model}
\usage{
stepper_components(from, to, R, K, r, bbox, dbg = 0, toroidal = 0)
}
\arguments{
\item{from}{coordinates of the data}

\item{to}{coordinate matrix for locations for evaluation of papangelou. Assumed to be different from 'from'}

\item{R}{maximum range}

\item{K}{the number of steps}

\item{r}{alternative to K and R, the r vector itself. Increasing, 0 shouldn't be included.}

\item{bbox}{bounding box}

\item{dbg}{dbg level}

\item{toroidal}{>0 use toroidal distances}
}
\value{
Matrix with dimension (nrow(to), K) with
d_k(to_i|from) = 1( ne_k(to_i, from)>0) with ne_k(x,from) = sum_j 1(r[k-1] < ||to_i - from_j|| < r[k])

If 'to' is NULL, computes the components for 'from' locations, i.e. the papangelou for each data point.
}
\description{
Compute the components of papangelou for the stepper model
}

