% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper-papangelou.R
\name{stepper_log_papangelou}
\alias{stepper_log_papangelou}
\title{Compute the papangelou conditional intensity for the stepper model}
\usage{
stepper_log_papangelou(from, to, theta, r, sat, bbox, dbg = 0, toroidal = 0)
}
\arguments{
\item{from}{coordinates of the data}

\item{to}{coordinate matrix for locations for evaluation of papangelou. Assumed to be different from 'from'}

\item{theta}{vector of weights, K+1 long}

\item{r}{steps, increasing, does not include 0}

\item{sat}{saturation level}

\item{bbox}{bounding box}

\item{dbg}{dbg level}

\item{toroidal}{>0 use toroidal distances}
}
\value{
log-conditional intensity l(to_i|from) = theta[0] + sum( theta[k] * d_k(to_i,from) )
where d_k(to_i|from) = 1( ne_k(to_i, from)>0) and ne_k(x,from) = sum_j 1(r[k-1] < ||to_i - from_j|| < r[k])
}
\description{
Compute the papangelou conditional intensity for the stepper model
}
