% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstepper-papangelou-components-multi-grid.R
\name{stepper_multi_components_grid}
\alias{stepper_multi_components_grid}
\title{Compute the components of papangelou for the stepper model, from a grid to data}
\usage{
stepper_multi_components_grid(from, to, ranges1, ranges2, sat1, sat2, bbox,
  dbg = 0, toroidal = 0, ...)
}
\arguments{
\item{from}{coordinates of the data}

\item{to}{coordinate matrix for locations for evaluation of papangelou. Assumed to be different from 'from'}

\item{bbox}{bounding box}

\item{dbg}{dbg level}

\item{toroidal}{>0 use toroidal distances}

\item{...}{ignored}

\item{K}{the number of steps}

\item{R}{maximum range}

\item{r}{alternative to K and R, the r vector itself. Increasing, 0 shouldn't be included.}

\item{sat}{the saturation levels, >0 integers.}
}
\value{
Matrix with dimension (nrow(to), K) with
d_k(to_i|from) = 1( ne_k(to_i, from)>0) with ne_k(x,from) = sum_j 1(r[k-1] < ||to_i - from_j|| < r[k])

If 'to' is NULL, computes the components for 'from' locations, i.e. the papangelou for each data point.
}
\description{
Will compute the components from all grid points to data, for each type of grid point.
}
