% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper-papangelou.R
\name{stepper_multi_log_papangelou}
\alias{stepper_multi_log_papangelou}
\title{Compute the papangelou conditional intensity for the multivariate stepper model}
\usage{
stepper_multi_log_papangelou(from, to, theta1, theta2, bbox, dbg = 0,
  toroidal = 0, multi_to = FALSE)
}
\arguments{
\item{from}{coordinates of the data, 3rd column 0-based integer marks}

\item{to}{coordinate matrix for locations for evaluation of papangelou, 0-based integer marks in 3rd column. Assumed to be different from 'from'. If NULL or missing, compute at "from".}

\item{theta1}{list of intra-type parameters, as in simulation (see rstepper_multi)}

\item{theta2}{list of inter-type parameters}

\item{bbox}{bounding box, defines the window.}

\item{dbg}{dbg level}

\item{toroidal}{>0 use toroidal distances}

\item{multi_to}{If TRUE, treat 'to' as unmarked locations at which to evaluate the papangelou for each mark level.}
}
\value{
log-conditional intensity l(to_i|from) = theta[0] + sum( theta[k] * d_k(to_i,from) )
where d_k(to_i|from) = 1( ne_k(to_i, from)>0) and ne_k(x,from) = sum_j 1(r[k-1] < ||to_i - from_j|| < r[k])

If multi_to = TRUE, a matrix with each column corresponding to each mark level.
}
\description{
Compute the papangelou conditional intensity for the multivariate stepper model
}
\details{
The first order effects (trends, baselines) are not included, make sure add them afterwards.
}
